pyproject.toml
[project]
name = "pratyayprofile-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncio>=4.0.0",
    "motor>=3.7.1",
    "pymongo>=4.16.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.2",
    "fastapi[standard]>=0.128.0",
]
